Moves helpers and makes test-constants
