+++ packages/sdk/test/client.test.ts
@@ -31,8 +31,9 @@ import {
   InvalidMaxMintableError,
 } from '../src/errors'
 import { MINTER_ROLE, NULL_ADDRESS, NON_NULL_ADDRESS, UINT32_MAX, NULL_BYTES32 } from '../src/utils/constants'
+import { DEFAULT_SALT, SOUND_FEE, ONE_HOUR, PRICE } from './test-constants'
 import { getSaltAsBytes32 } from '../src/utils/helpers'
-import { MerkleTestHelper, now } from './helpers'
+import { MerkleTestHelper, now, getGenericEditionConfig, getGenericRangeMintConfig } from './helpers'
 
 import type { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
 import type MerkleTree from 'merkletreejs'
@@ -40,11 +41,6 @@ import type MerkleTree from 'merkletreejs'
 import type { ContractCall, MintConfig, MintSchedule } from '../src/types'
 import { MockAPI } from './helpers/api'
 import { randomUUID } from 'crypto'
-
-const DEFAULT_SALT = getSaltAsBytes32(12345678)
-const SOUND_FEE = 0
-const ONE_HOUR = 3600
-const PRICE = 420420420
 
 const SoundCreatorV1 = new SoundCreatorV1__factory()
 const SoundFeeRegistry = new SoundFeeRegistry__factory()
@@ -969,38 +965,6 @@ describe('mint', () => {
 
 describe('createEdition', () => {
   const SALT = 'hello'
-  const getGenericEditionConfig = () => ({
-    name: 'Test',
-    symbol: 'TEST',
-    metadataModule: NULL_ADDRESS,
-    baseURI: 'https://test.com',
-    contractURI: 'https://test.com',
-    fundingRecipient: NON_NULL_ADDRESS,
-    royaltyBPS: 0,
-    editionMaxMintableLower: 10,
-    editionMaxMintableUpper: 10,
-    editionCutoffTime: 999999,
-    shouldEnableMintRandomness: true,
-    shouldFreezeMetadata: false,
-    enableOperatorFiltering: true,
-  })
-
-  const startTime = now()
-  const cutoffTime = startTime + ONE_HOUR / 2
-  const endTime = cutoffTime + ONE_HOUR
-
-  const getGenericRangeMintConfig = () => ({
-    mintType: 'RangeEdition' as const,
-    minterAddress: rangeEditionMinter.address,
-    price: PRICE,
-    startTime,
-    cutoffTime,
-    endTime,
-    maxMintableLower: 3,
-    maxMintableUpper: 4,
-    maxMintablePerAccount: 1,
-    affiliateFeeBPS: 0,
-  })
 
   beforeEach(() => {
     client = SoundClient({ signer: artistWallet, soundCreatorAddress: soundCreator.address })
@@ -1122,7 +1086,7 @@ describe('createEdition', () => {
     await client
       .createEdition({
         editionConfig,
-        mintConfigs: [getGenericRangeMintConfig()],
+        mintConfigs: [getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })],
         salt: SALT,
       })
       .catch((error) => {
@@ -1138,7 +1102,7 @@ describe('createEdition', () => {
     await client
       .createEdition({
         editionConfig,
-        mintConfigs: [getGenericRangeMintConfig()],
+        mintConfigs: [getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })],
         salt: SALT,
       })
       .catch((error) => {
@@ -1149,7 +1113,7 @@ describe('createEdition', () => {
   it('throws if maxMintablePerAccount is zero', async () => {
     const editionConfig = getGenericEditionConfig()
 
-    const mintConfig = getGenericRangeMintConfig()
+    const mintConfig = getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })
     mintConfig.maxMintablePerAccount = 0
 
     await client
@@ -1166,7 +1130,7 @@ describe('createEdition', () => {
   it('throws if range mint maxMintableLower exceeds maxMintableUpper', async () => {
     const editionConfig = getGenericEditionConfig()
 
-    const mintConfig = getGenericRangeMintConfig()
+    const mintConfig = getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })
     mintConfig.maxMintableLower = 2
     mintConfig.maxMintableUpper = 1
 
@@ -1184,7 +1148,7 @@ describe('createEdition', () => {
   it('throws if range mint startTime == cutoffTime', async () => {
     const editionConfig = getGenericEditionConfig()
 
-    const mintConfig = getGenericRangeMintConfig()
+    const mintConfig = getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })
     mintConfig.startTime = mintConfig.cutoffTime
 
     await client
@@ -1201,7 +1165,7 @@ describe('createEdition', () => {
   it('throws if range mint cutoffTime === endTime', async () => {
     const editionConfig = getGenericEditionConfig()
 
-    const mintConfig = getGenericRangeMintConfig()
+    const mintConfig = getGenericRangeMintConfig({ minterAddress: rangeEditionMinter.address })
     mintConfig.cutoffTime = mintConfig.endTime
 
     await client
