+++ packages/sdk/test/test-constants.ts
@@ -1,0 +1,6 @@
+import { getSaltAsBytes32 } from '../dist/utils/helpers'
+
+export const DEFAULT_SALT = getSaltAsBytes32(12345678)
+export const SOUND_FEE = 0
+export const ONE_HOUR = 3600
+export const PRICE = 420420420
